Patch/food pack20.12.25.2

+++ src/main/java/org/acme/foodpackaging/dto/JobSelectionDTO.java
@@ -3,6 +3,6 @@ package org.acme.foodpackaging.dto;
 import java.util.Map;
 
 public record JobSelectionDTO(
-        Map<Integer, Boolean> selection
+        Map<Long, Boolean> selection
 ) {}
