+++ src/main/java/org/acme/foodpackaging/factory/JobFactory.java
@@ -15,7 +15,7 @@ public class JobFactory {
     @Inject
     LoadDataService loadDataService;
 
-    public Job createJob(String id, String lineId, int snpz, int np, String jobName, Product product, double mass,
+    public Job createJob(String id, String lineId, long snpz, int np, String jobName, Product product, double mass,
             int quantity, int duration, LocalDateTime minStartDateTime,
                          LocalDateTime idealEndDateTime,
                          LocalDateTime maxEndDateTime,  int priority, LocalDateTime startProductionDateTime) {
